moved help section to bottom of sidebar, refactoring, notes
moved help section to bottom of sidebar, refactoring, notes
