+++ templates/interface/php/admin/admin-login/login.php
@@ -34,7 +34,7 @@ if ( $_POST['admin_submit_login'] ) {
 
 
 $login_template = 1;
-require("templates/interface/php/header.php");
+require("templates/interface/php/wrap/header.php");
 
 ?>
 
@@ -203,5 +203,5 @@ Google Fonts is supported (fonts.google.com).'>Get A Different Image</a>
 
 
 <?php
-require("templates/interface/php/footer.php");
+require("templates/interface/php/wrap/footer.php");
 ?>
